move fixtures to common conftest
